==> aou_workbench_siloed_analyses/03_aou_variant_qc.ipynb <==
# Perform variant quality control

In this notebook, we perform variant quality control using [PLINK2](https://www.cog-genomics.org/plink/2.0/).

Note that this work is part of a larger project to [Demonstrate the Potential for Pooled Analysis of All of Us and UK Biobank Genomic Data](https://github.com/all-of-us/ukb-cross-analysis-demo-project). Specifically this is for the portion of the project that is the **siloed** analysis.

# Setup 

<div class="alert alert-block alert-warning">
    <b>Cloud Environment</b>: This notebook was written for use on the <i>All of Us</i> Workbench.
    <ul>
        <li>Use "Recommended Environment" <kbd><b>General Analysis</b></kbd> which creates compute type <kbd><b>Standard VM</b></kbd> and change the resources to <b>8 CPUs</b> and <b>30 GB RAM</b>.</li>
        <li>This notebook will also run fine with the default values for CPU and RAM but the subsequent notebooks (e.g., PCA and regenie) will benefit from a machine with more resources.</li>
        <li>When run on a standard VM with resources 8 CPUs and 30 GB RAM, this notebook takes only 10 minutes to run. You can also run it in the background via <kbd>run_notebook_in_the_background</kbd> for the sake of provenance and reproducibility.</li>
    </ul>
</div>

from datetime import datetime
import os
import time

## Setup plink2

https://www.cog-genomics.org/plink/2.0/

%%bash

##### plink 2 install
PLINK_VERSION=2.3.Alpha
PLINK_ZIP_PATH=/tmp/plink-$PLINK_VERSION.zip
curl -L -o $PLINK_ZIP_PATH https://s3.amazonaws.com/plink2-assets/alpha2/plink2_linux_x86_64.zip
mkdir -p /tmp/plink2/
unzip -o $PLINK_ZIP_PATH -d /tmp/plink2/

!/tmp/plink2/plink2 --version # --help

## Define constants

# Papermill parameters. See https://papermill.readthedocs.io/en/latest/usage-parameterize.html

#---[ Inputs ]---
# The BGEN file was created via aou_workbench_siloed_analyses/02_aou_write_filtered_bgen.ipynb.
REMOTE_BGEN = 'gs://fc-secure-098ff3db-05c2-4426-8914-a26608668529/data/aou/geno/20220315/aou-alpha3-chr1-chr22.bgen'
# The sample file was created via aou_workbench_siloed_analyses/02_aou_write_filtered_bgen.ipynb.
REMOTE_BGEN_SAMPLE = 'gs://fc-secure-098ff3db-05c2-4426-8914-a26608668529/data/aou/geno/20220315/aou-alpha3-chr1-chr22.sample'
# This TSV was created via notebook aou_workbench_siloed_analyses/01_aou_lipids_phenotype.ipynb
REMOTE_PHENOTYPE_SAMPLE_IDS = 'gs://fc-secure-471c1068-cd3d-4b43-9b5d-a618c85ceea5/data/aou/pheno/20220311/aou_alpha3_lipids_ids.tsv'

#---[ Outputs ]---
# Create a timestamp for a folder of results generated today.
DATESTAMP = time.strftime('%Y%m%d')
OUTPUT_FILENAME_PREFIX = 'aou_alpha3_lipids'
OUTPUT_FOLDER = f'{os.getenv("WORKSPACE_BUCKET")}/data/aou/variant-qc/{DATESTAMP}/'

LOCAL_BGEN = os.path.basename(REMOTE_BGEN)
LOCAL_BGEN_SAMPLE = os.path.basename(REMOTE_BGEN_SAMPLE)
LOCAL_PHENOTYPE_SAMPLE_IDS = os.path.basename(REMOTE_PHENOTYPE_SAMPLE_IDS)

## Copy data locally

!gsutil cp -n {REMOTE_BGEN} {REMOTE_BGEN_SAMPLE} .

!gsutil cp {REMOTE_PHENOTYPE_SAMPLE_IDS} .

# Variant QC for regenie step 1 via plink2

Use plink2 to perform [variant QC appropriate for step 1 of regenie](https://rgcgithub.github.io/regenie/recommendations/#exclusion-files), obtaining a subset of high quality variants roughly equal to the number of samples.

!/tmp/plink2/plink2 \
  --bgen {LOCAL_BGEN} ref-first \
  --sample {LOCAL_BGEN_SAMPLE} \
  --chr 1-22 \
  --keep {LOCAL_PHENOTYPE_SAMPLE_IDS} \
  --geno 0.1 \
  --mind 0.1 \
  --mac 100 \
  --hwe 1e-15 \
  --write-snplist \
  --write-samples \
  --no-id-header \
  --out {OUTPUT_FILENAME_PREFIX}_step1QC_plink

!ls -lth . | head

!head {OUTPUT_FILENAME_PREFIX}_step1QC_plink.id

!wc -l {OUTPUT_FILENAME_PREFIX}_step1QC_plink.id

!head {OUTPUT_FILENAME_PREFIX}_step1QC_plink.snplist

!wc -l {OUTPUT_FILENAME_PREFIX}_step1QC_plink.snplist

!ls {OUTPUT_FILENAME_PREFIX}_step1QC_plink*

!gsutil -m cp {OUTPUT_FILENAME_PREFIX}_step1QC_plink* {OUTPUT_FOLDER}

# Variant QC for regenie step 2 via plink2

Use plink2 to perform [variant QC appropriate for step 2 of regenie](https://rgcgithub.github.io/regenie/recommendations/#exclusion-files), obtaining a subset of high quality variants.

!/tmp/plink2/plink2 \
  --bgen {LOCAL_BGEN} ref-first \
  --sample {LOCAL_BGEN_SAMPLE} \
  --chr 1-22 \
  --keep {LOCAL_PHENOTYPE_SAMPLE_IDS} \
  --geno 0.1 \
  --mind 0.1 \
  --mac 6 \
  --hwe 1e-15 \
  --write-snplist \
  --write-samples \
  --no-id-header \
  --out {OUTPUT_FILENAME_PREFIX}_step2QC_plink

# This differs from step 1 QC in that the following parameter was changed
#    --mac 100

!ls -lth . | head

!head {OUTPUT_FILENAME_PREFIX}_step2QC_plink.id

!wc -l {OUTPUT_FILENAME_PREFIX}_step2QC_plink.id

!head {OUTPUT_FILENAME_PREFIX}_step2QC_plink.snplist

!wc -l {OUTPUT_FILENAME_PREFIX}_step2QC_plink.snplist

!ls {OUTPUT_FILENAME_PREFIX}_step2QC_plink*

!gsutil -m cp {OUTPUT_FILENAME_PREFIX}_step2QC_plink* {OUTPUT_FOLDER}

!gsutil ls {OUTPUT_FOLDER}

# Provenance 

%%bash

date

%%bash

pip3 freeze
