==> aou_workbench_siloed_analyses/6_aggregate_gwas_results.ipynb <==
# Aggregate GWAS results

In this notebook we modify our GWAS results such that they become okay to leave the *All of Us* Researcher workbench per the data dissemination rules.

# Setup

<div class="alert alert-block alert-warning">
    <b>Cloud Environment</b>: This notebook was written for use on All of Us Workbench. It runs fine on the default Cloud Environment. 
</div>

library(tidyverse)

# Constants

# Papermill parameters. See https://papermill.readthedocs.io/en/latest/usage-parameterize.html

REGENIE_RESULTS <- c(
    HDL='gs://fc-secure-fd6786bf-6c28-4f33-ac30-3860fbeee5bb/data/regenie/20210921/aou_alpha2_ukb_lipids_regenie_part2_HDLnorm.regenie',
    LDL='gs://fc-secure-fd6786bf-6c28-4f33-ac30-3860fbeee5bb/data/regenie/20210921/aou_alpha2_ukb_lipids_regenie_part2_LDLnorm.regenie',
    TC='gs://fc-secure-fd6786bf-6c28-4f33-ac30-3860fbeee5bb/data/regenie/20210921/aou_alpha2_ukb_lipids_regenie_part2_TCnorm.regenie',
    TG='gs://fc-secure-fd6786bf-6c28-4f33-ac30-3860fbeee5bb/data/regenie/20210921/aou_alpha2_ukb_lipids_regenie_part2_TGnorm.regenie'
)

LIPIDS <- names(REGENIE_RESULTS)

RAW_FILE_SUFFIX <- '.regenie'
AGGREGATE_FILE_SUFFIX <- '_aggregated.tsv'

# Load the regenie GWAS results

Bring our results into a single dataframe with a lipid type column.

combined_regenie_results <- bind_rows(
    lapply(LIPIDS, function(lipid) {
        file = REGENIE_RESULTS[lipid]
        read_delim(pipe(str_glue('gsutil cat {file}')), delim = ' ') %>%
        mutate(lipid_type = lipid)
    })) %>%
    mutate(
        AN = 2 * N,
        AC_alt = A1FREQ * AN,
        AC_ref = (1 - A1FREQ) * AN
    )

dim(combined_regenie_results)

head(combined_regenie_results)

combined_regenie_results %>%
    group_by(lipid_type) %>%
    summarize(
        count = n(),
        min_LOG10P = min(LOG10P),
        max_LOG10P = max(LOG10P),
        min_A1FREQ = min(A1FREQ),
        max_A1FREQ = max(A1FREQ),
        min_N = min(N),
        max_N = max(N),
        min_AC_alt = min(AC_alt),
        max_AC_alt = max(AC_alt),
        min_AC_ref = min(AC_ref),
        max_AC_ref = max(AC_ref),
    )

## How many significant results will be removed from the aggregate?

combined_regenie_results %>%
    mutate(significant = LOG10P > -log10(5e-08)) %>%
    group_by(lipid_type, significant) %>%
    summarize(count = n())

# Filter to ensure at least 20 individuals have the variant

aggregate_regenie_results <- combined_regenie_results %>%
    filter(AC_alt >= 40 & AC_ref >= 40)

aggregate_regenie_results %>%
    group_by(lipid_type) %>%
    summarize(
        count = n(),
        min_LOG10P = min(LOG10P),
        max_LOG10P = max(LOG10P),
        min_A1FREQ = min(A1FREQ),
        max_A1FREQ = max(A1FREQ),
        min_N = min(N),
        max_N = max(N),
        min_AC_alt = min(AC_alt),
        max_AC_alt = max(AC_alt),
        min_AC_ref = min(AC_ref),
        max_AC_ref = max(AC_ref),
    )

# Write out the aggregate data to local disk

for (lipid in LIPIDS) {
    input_file <- REGENIE_RESULTS[lipid]
    output_file <- input_file %>%
        str_replace('gs://', '') %>%
        str_replace_all('/', '_') %>%
        str_replace(str_glue('{RAW_FILE_SUFFIX}$'), AGGREGATE_FILE_SUFFIX)
    message(str_glue('Aggregating results from {input_file} to {output_file}'))
    stopifnot('output filename must be different from input filename' =
              output_file != input_file)
    write_tsv(aggregate_regenie_results %>% filter(lipid_type == lipid), output_file)
}

# Now you can download these files!

**Be sure to download the aggregated TSV files**, not the .regenie files with the raw results.

# Provenance 

devtools::session_info()
